defects fixed
