+++ src/main/java/ru/mail/polis/dao/basta123/RecordIter.java
@@ -9,13 +9,12 @@ import java.util.Iterator;
 import static ru.mail.polis.service.basta123.Utils.getByteArrayFromByteBuffer;
 import static ru.mail.polis.service.basta123.Utils.getByteBufferFromByteArray;
 
-public class MyRecordIter implements Iterator<Record> {
+public class RecordIter implements Iterator<Record> {
 
     private final RocksIterator rocksIterator;
 
-    public MyRecordIter(final ByteBuffer key, final RocksIterator rocksIterator) {
+    public RecordIter(final RocksIterator rocksIterator) {
         this.rocksIterator = rocksIterator;
-        this.rocksIterator.seek(getByteArrayFromByteBuffer(key));
     }
 
     @Override
@@ -25,6 +24,9 @@ public class MyRecordIter implements Iterator<Record> {
 
     @Override
     public Record next() {
+        if (!hasNext()) {
+            throw new IllegalStateException("next doesn't exist");
+        }
         final byte[] keyBytes = rocksIterator.key();
         final byte[] valueBytes = rocksIterator.value();
         rocksIterator.next();
