+++ src/main/java/ru/mail/polis/dao/basta123/SingedBytesComparator.java
@@ -5,9 +5,9 @@ import org.rocksdb.Comparator;
 import org.rocksdb.ComparatorOptions;
 import org.rocksdb.Slice;
 
-public class MyComparator extends Comparator {
+public class SingedBytesComparator extends Comparator {
 
-    public MyComparator(final ComparatorOptions compOptions) {
+    public SingedBytesComparator(final ComparatorOptions compOptions) {
         super(compOptions);
     }
